requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==4.2.16
django-crispy-forms==2.3
mysqlclient==2.2.4
pillow==10.4.0
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
